requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
pypdf